requirements.txt
Django==2.2.24
djangorestframework==3.12.4
drf-writable-nested==0.6.4
drf-yasg==1.21.10
pytest==7.4.4
pytest-django==4.5.2